Fix test reorganization
