+++ tests/Functional/Apps/DefaultApp/src/Controller/DefaultRowAction/DefaultRowActionDetailCrudController.php
@@ -1,21 +1,20 @@
 <?php
 
-namespace EasyCorp\Bundle\EasyAdminBundle\Tests\TestApplication\Controller;
+namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Functional\Apps\DefaultApp\Controller\DefaultRowAction;
 
 use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
-use EasyCorp\Bundle\EasyAdminBundle\Tests\TestApplication\Entity\Category;
+use EasyCorp\Bundle\EasyAdminBundle\Tests\Functional\Apps\DefaultApp\Entity\Category;
 
 /**
- * Tests the default row action fallback behavior.
- * EDIT is disabled, DETAIL is enabled, so the default row action should fallback to DETAIL.
+ * Tests the setDefaultRowAction() method with Action::DETAIL.
  *
  * @extends AbstractCrudController<Category>
  */
-class DefaultRowActionFallbackCrudController extends AbstractCrudController
+class DefaultRowActionDetailCrudController extends AbstractCrudController
 {
     public static function getEntityFqcn(): string
     {
@@ -29,10 +28,14 @@ class DefaultRowActionFallbackCrudController extends AbstractCrudController
         ];
     }
 
+    public function configureCrud(Crud $crud): Crud
+    {
+        return $crud->setDefaultRowAction(Action::DETAIL);
+    }
+
     public function configureActions(Actions $actions): Actions
     {
         return $actions
-            ->disable(Action::EDIT)
             ->add(Crud::PAGE_INDEX, Action::DETAIL)
         ;
     }
