+++ tests/Functional/Apps/DefaultApp/src/Controller/DefaultRowAction/DefaultRowActionFallbackCrudController.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace EasyCorp\Bundle\EasyAdminBundle\Tests\TestApplication\Controller;
+namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Functional\Apps\DefaultApp\Controller\DefaultRowAction;
 
 use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
-use EasyCorp\Bundle\EasyAdminBundle\Tests\TestApplication\Entity\Category;
+use EasyCorp\Bundle\EasyAdminBundle\Tests\Functional\Apps\DefaultApp\Entity\Category;
 
 /**
- * Tests the setDefaultRowAction() method with null (disabled).
+ * Tests the default row action fallback behavior.
+ * EDIT is disabled, DETAIL is enabled, so the default row action should fallback to DETAIL.
  *
  * @extends AbstractCrudController<Category>
  */
-class DefaultRowActionDisabledCrudController extends AbstractCrudController
+class DefaultRowActionFallbackCrudController extends AbstractCrudController
 {
     public static function getEntityFqcn(): string
     {
@@ -28,14 +29,10 @@ class DefaultRowActionDisabledCrudController extends AbstractCrudController
         ];
     }
 
-    public function configureCrud(Crud $crud): Crud
-    {
-        return $crud->setDefaultRowAction(null);
-    }
-
     public function configureActions(Actions $actions): Actions
     {
         return $actions
+            ->disable(Action::EDIT)
             ->add(Crud::PAGE_INDEX, Action::DETAIL)
         ;
     }
