+++ tests/Unit/Filter/Configurator/CurrencyConfiguratorTest.php
@@ -1,28 +1,28 @@
 <?php
 
-namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Filter\Configurator;
+namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Unit\Filter\Configurator;
 
 use Doctrine\ORM\Mapping\ClassMetadata;
 use EasyCorp\Bundle\EasyAdminBundle\Context\AdminContext;
 use EasyCorp\Bundle\EasyAdminBundle\Context\I18nContext;
 use EasyCorp\Bundle\EasyAdminBundle\Context\RequestContext;
 use EasyCorp\Bundle\EasyAdminBundle\Dto\EntityDto;
-use EasyCorp\Bundle\EasyAdminBundle\Filter\Configurator\CountryConfigurator;
-use EasyCorp\Bundle\EasyAdminBundle\Filter\CountryFilter;
+use EasyCorp\Bundle\EasyAdminBundle\Filter\Configurator\CurrencyConfigurator;
+use EasyCorp\Bundle\EasyAdminBundle\Filter\CurrencyFilter;
 use EasyCorp\Bundle\EasyAdminBundle\Filter\TextFilter;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Intl\Countries;
+use Symfony\Component\Intl\Currencies;
 
-class CountryConfiguratorTest extends TestCase
+class CurrencyConfiguratorTest extends TestCase
 {
-    private CountryConfigurator $configurator;
+    private CurrencyConfigurator $configurator;
     private EntityDto $entityDto;
     private AdminContext $adminContext;
 
     protected function setUp(): void
     {
-        $this->configurator = new CountryConfigurator();
+        $this->configurator = new CurrencyConfigurator();
 
         $metadata = new ClassMetadata('App\Entity\Test');
         $metadata->setIdentifier(['id']);
@@ -36,9 +36,9 @@ class CountryConfiguratorTest extends TestCase
         );
     }
 
-    public function testSupportsCountryFilter(): void
+    public function testSupportsCurrencyFilter(): void
     {
-        $filter = CountryFilter::new('country');
+        $filter = CurrencyFilter::new('currency');
         $filterDto = $filter->getAsDto();
 
         $this->assertTrue($this->configurator->supports($filterDto, null, $this->entityDto, $this->adminContext));
@@ -52,23 +52,23 @@ class CountryConfiguratorTest extends TestCase
         $this->assertFalse($this->configurator->supports($filterDto, null, $this->entityDto, $this->adminContext));
     }
 
-    public function testConfigureSetsAllCountries(): void
+    public function testConfigureSetsAllCurrencies(): void
     {
-        $filter = CountryFilter::new('country');
+        $filter = CurrencyFilter::new('currency');
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
-        $expectedCountries = array_flip(Countries::getNames());
+        $expectedCurrencies = array_flip(Currencies::getNames());
 
-        $this->assertSame($expectedCountries, $choices);
+        $this->assertSame($expectedCurrencies, $choices);
         $this->assertFalse($filterDto->getFormTypeOption('value_type_options.choice_translation_domain'));
     }
 
     public function testConfigureWithIncludeOnly(): void
     {
-        $filter = CountryFilter::new('country')->includeOnly(['ES', 'FR', 'DE']);
+        $filter = CurrencyFilter::new('currency')->includeOnly(['EUR', 'USD', 'GBP']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
@@ -76,49 +76,35 @@ class CountryConfiguratorTest extends TestCase
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
 
         $this->assertCount(3, $choices);
-        $this->assertArrayHasKey(Countries::getName('ES'), $choices);
-        $this->assertArrayHasKey(Countries::getName('FR'), $choices);
-        $this->assertArrayHasKey(Countries::getName('DE'), $choices);
-        $this->assertSame('ES', $choices[Countries::getName('ES')]);
+        $this->assertArrayHasKey(Currencies::getName('EUR'), $choices);
+        $this->assertArrayHasKey(Currencies::getName('USD'), $choices);
+        $this->assertArrayHasKey(Currencies::getName('GBP'), $choices);
+        $this->assertSame('EUR', $choices[Currencies::getName('EUR')]);
     }
 
     public function testConfigureWithRemove(): void
     {
-        $filter = CountryFilter::new('country')->remove(['US', 'CA']);
+        $filter = CurrencyFilter::new('currency')->remove(['EUR', 'USD']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
 
-        $this->assertArrayNotHasKey(Countries::getName('US'), $choices);
-        $this->assertArrayNotHasKey(Countries::getName('CA'), $choices);
-        $this->assertArrayHasKey(Countries::getName('ES'), $choices);
+        $this->assertArrayNotHasKey(Currencies::getName('EUR'), $choices);
+        $this->assertArrayNotHasKey(Currencies::getName('USD'), $choices);
+        $this->assertArrayHasKey(Currencies::getName('GBP'), $choices);
     }
 
     public function testConfigureWithPreferredChoices(): void
     {
-        $filter = CountryFilter::new('country')->preferredChoices(['ES', 'FR']);
+        $filter = CurrencyFilter::new('currency')->preferredChoices(['EUR', 'USD']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $preferredChoices = $filterDto->getFormTypeOption('value_type_options.preferred_choices');
 
-        $this->assertSame(['ES', 'FR'], $preferredChoices);
-    }
-
-    public function testConfigureWithAlpha3Codes(): void
-    {
-        $filter = CountryFilter::new('country')->useAlpha3Codes()->includeOnly(['ESP', 'FRA']);
-        $filterDto = $filter->getAsDto();
-
-        $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
-
-        $choices = $filterDto->getFormTypeOption('value_type_options.choices');
-
-        $this->assertCount(2, $choices);
-        $this->assertSame('ESP', $choices[Countries::getAlpha3Name('ESP')]);
-        $this->assertSame('FRA', $choices[Countries::getAlpha3Name('FRA')]);
+        $this->assertSame(['EUR', 'USD'], $preferredChoices);
     }
 }
