+++ tests/Unit/Filter/Configurator/LocaleConfiguratorTest.php
@@ -1,28 +1,28 @@
 <?php
 
-namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Filter\Configurator;
+namespace EasyCorp\Bundle\EasyAdminBundle\Tests\Unit\Filter\Configurator;
 
 use Doctrine\ORM\Mapping\ClassMetadata;
 use EasyCorp\Bundle\EasyAdminBundle\Context\AdminContext;
 use EasyCorp\Bundle\EasyAdminBundle\Context\I18nContext;
 use EasyCorp\Bundle\EasyAdminBundle\Context\RequestContext;
 use EasyCorp\Bundle\EasyAdminBundle\Dto\EntityDto;
-use EasyCorp\Bundle\EasyAdminBundle\Filter\Configurator\CurrencyConfigurator;
-use EasyCorp\Bundle\EasyAdminBundle\Filter\CurrencyFilter;
+use EasyCorp\Bundle\EasyAdminBundle\Filter\Configurator\LocaleConfigurator;
+use EasyCorp\Bundle\EasyAdminBundle\Filter\LocaleFilter;
 use EasyCorp\Bundle\EasyAdminBundle\Filter\TextFilter;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Intl\Currencies;
+use Symfony\Component\Intl\Locales;
 
-class CurrencyConfiguratorTest extends TestCase
+class LocaleConfiguratorTest extends TestCase
 {
-    private CurrencyConfigurator $configurator;
+    private LocaleConfigurator $configurator;
     private EntityDto $entityDto;
     private AdminContext $adminContext;
 
     protected function setUp(): void
     {
-        $this->configurator = new CurrencyConfigurator();
+        $this->configurator = new LocaleConfigurator();
 
         $metadata = new ClassMetadata('App\Entity\Test');
         $metadata->setIdentifier(['id']);
@@ -36,9 +36,9 @@ class CurrencyConfiguratorTest extends TestCase
         );
     }
 
-    public function testSupportsCurrencyFilter(): void
+    public function testSupportsLocaleFilter(): void
     {
-        $filter = CurrencyFilter::new('currency');
+        $filter = LocaleFilter::new('locale');
         $filterDto = $filter->getAsDto();
 
         $this->assertTrue($this->configurator->supports($filterDto, null, $this->entityDto, $this->adminContext));
@@ -52,23 +52,23 @@ class CurrencyConfiguratorTest extends TestCase
         $this->assertFalse($this->configurator->supports($filterDto, null, $this->entityDto, $this->adminContext));
     }
 
-    public function testConfigureSetsAllCurrencies(): void
+    public function testConfigureSetsAllLocales(): void
     {
-        $filter = CurrencyFilter::new('currency');
+        $filter = LocaleFilter::new('locale');
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
-        $expectedCurrencies = array_flip(Currencies::getNames());
+        $expectedLocales = array_flip(Locales::getNames());
 
-        $this->assertSame($expectedCurrencies, $choices);
+        $this->assertSame($expectedLocales, $choices);
         $this->assertFalse($filterDto->getFormTypeOption('value_type_options.choice_translation_domain'));
     }
 
     public function testConfigureWithIncludeOnly(): void
     {
-        $filter = CurrencyFilter::new('currency')->includeOnly(['EUR', 'USD', 'GBP']);
+        $filter = LocaleFilter::new('locale')->includeOnly(['en_US', 'es_ES', 'fr_FR']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
@@ -76,35 +76,35 @@ class CurrencyConfiguratorTest extends TestCase
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
 
         $this->assertCount(3, $choices);
-        $this->assertArrayHasKey(Currencies::getName('EUR'), $choices);
-        $this->assertArrayHasKey(Currencies::getName('USD'), $choices);
-        $this->assertArrayHasKey(Currencies::getName('GBP'), $choices);
-        $this->assertSame('EUR', $choices[Currencies::getName('EUR')]);
+        $this->assertArrayHasKey(Locales::getName('en_US'), $choices);
+        $this->assertArrayHasKey(Locales::getName('es_ES'), $choices);
+        $this->assertArrayHasKey(Locales::getName('fr_FR'), $choices);
+        $this->assertSame('en_US', $choices[Locales::getName('en_US')]);
     }
 
     public function testConfigureWithRemove(): void
     {
-        $filter = CurrencyFilter::new('currency')->remove(['EUR', 'USD']);
+        $filter = LocaleFilter::new('locale')->remove(['en_US', 'es_ES']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $choices = $filterDto->getFormTypeOption('value_type_options.choices');
 
-        $this->assertArrayNotHasKey(Currencies::getName('EUR'), $choices);
-        $this->assertArrayNotHasKey(Currencies::getName('USD'), $choices);
-        $this->assertArrayHasKey(Currencies::getName('GBP'), $choices);
+        $this->assertArrayNotHasKey(Locales::getName('en_US'), $choices);
+        $this->assertArrayNotHasKey(Locales::getName('es_ES'), $choices);
+        $this->assertArrayHasKey(Locales::getName('fr_FR'), $choices);
     }
 
     public function testConfigureWithPreferredChoices(): void
     {
-        $filter = CurrencyFilter::new('currency')->preferredChoices(['EUR', 'USD']);
+        $filter = LocaleFilter::new('locale')->preferredChoices(['en_US', 'es_ES']);
         $filterDto = $filter->getAsDto();
 
         $this->configurator->configure($filterDto, null, $this->entityDto, $this->adminContext);
 
         $preferredChoices = $filterDto->getFormTypeOption('value_type_options.preferred_choices');
 
-        $this->assertSame(['EUR', 'USD'], $preferredChoices);
+        $this->assertSame(['en_US', 'es_ES'], $preferredChoices);
     }
 }
